add openNlp as a NLP library to tokenize the input document content  avoiding stop words of the document corresponding language (using fr and en languages for the moment)

+++ src/main/java/com/maltem/aiboo/batch/model/ConceptOccurence.java
@@ -28,7 +28,7 @@ public class ConceptOccurence implements Serializable {
     @JoinColumn(name = "concept_name")
     private Concept concept;
 
-    private int occurence;
+    private Long occurence;
 
     /*public ConceptOccurence(Concept concept, int occurence) {
         this.concept = concept;
